Migrated tests from test_helper.py into test_csv_utils.py
